S9/DogsBreedsClassifier/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
# --------- pytorch --------- #
timm==1.0.11
torch==2.4.0
torchvision==0.19.0
lightning==2.4.0
torchmetrics==1.4.1
tensorboard==2.17.1
tensorboardX==2.6.2.2

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
# hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
# pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
# pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

# --------- deployment --------- #
gradio==3.3.1

# ---------- pytest ------------- #
pytest==8.3.3 
pytest-cov==5.0.0

rootutils
loguru
scikit-image==0.22.0
scikit-learn==1.3.1
hydra-optuna-sweeper==1.2.0
optuna==2.10.1
hydra-joblib-launcher==1.2.0
joblib==1.3.2
dvc==3.55.2
dvc-data==3.16.6
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.40.1
